+++ assets/images/12-sql-homework-demo-01.png
