+++ assets/images/12-sql-homework-video-thumbnail.png
